unity_test/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.7.0
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
zipp==3.8.1
